avatar/interactive/requirements.txt
python-dotenv
retry
azure-core
azure-search-documents==11.4.0b6
openai==0.28.1
tenacity
pyodbc